Adds docker-compose deploy for Prefect flows
Extends the Makefile with commands to deploy Prefect flows using docker-compose.
This allows users to easily deploy and manage flows within a Docker environment
by specifying a separate configuration file (prefect.docker.yaml).

Specifically, it adds targets for:
- Deploying flows via `prefect deploy` with a configurable prefect file
- Setting the image tag specifically for the docker deployment
- Setting the environment name specifically for the docker deployment
- Setting the config for the docker deployment
- Deploying flows with a full set of commands, specific to docker

Removes the setting of the API URL, and INAGE_NAME from the configuration,
and sets a default tag for the IMAGE.

The default Prefect configuration is now handled by the `prefect.yaml` file.